refactor: move modals to separate module
