+++ core-commonui/modals/build.gradle.kts
@@ -15,50 +15,33 @@ kotlin {
             }
         }
     }
-    listOf(
-        iosX64(),
-        iosArm64(),
-        iosSimulatorArm64()
-    ).forEach {
-        it.binaries.framework {
-            baseName = "feature-inbox"
-        }
-    }
+    iosX64()
+    iosArm64()
+    iosSimulatorArm64()
 
     sourceSets {
         val commonMain by getting {
             dependencies {
-                implementation(libs.koin.core)
-
                 implementation(compose.runtime)
                 implementation(compose.foundation)
                 implementation(compose.material)
                 implementation(compose.material3)
-                implementation(compose.materialIconsExtended)
                 @OptIn(org.jetbrains.compose.ExperimentalComposeLibrary::class)
                 implementation(compose.components.resources)
+                implementation(compose.materialIconsExtended)
 
+                implementation(libs.koin.core)
                 implementation(libs.voyager.navigator)
-                implementation(libs.voyager.screenmodel)
-                implementation(libs.voyager.tab)
                 implementation(libs.voyager.bottomsheet)
 
-                implementation(projects.resources)
-                implementation(projects.coreArchitecture)
-                implementation(projects.coreAppearance)
-                implementation(projects.coreNavigation)
-                implementation(projects.coreCommonui)
-                implementation(projects.coreCommonui.components)
-                implementation(projects.coreCommonui.lemmyui)
-                implementation(projects.coreCommonui.modals)
                 implementation(projects.coreUtils)
-                implementation(projects.corePreferences)
+                implementation(projects.coreAppearance)
+                implementation(projects.coreCommonui.components)
+                implementation(projects.coreNavigation)
                 implementation(projects.corePersistence)
                 implementation(projects.coreNotifications)
-
                 implementation(projects.domainLemmy.data)
-                implementation(projects.domainLemmy.repository)
-                implementation(projects.domainIdentity)
+                implementation(projects.resources)
             }
         }
         val commonTest by getting {
@@ -70,7 +53,7 @@ kotlin {
 }
 
 android {
-    namespace = "com.github.diegoberaldin.raccoonforlemmy.feature.inbox"
+    namespace = "com.github.diegoberaldin.raccoonforlemmy.core.commonui.modals"
     compileSdk = libs.versions.android.targetSdk.get().toInt()
     defaultConfig {
         minSdk = libs.versions.android.minSdk.get().toInt()
