+++ core-commonui/modals/src/androidMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/core/commonui/modals/di/Utils.kt
@@ -1,0 +1,18 @@
+package com.github.diegoberaldin.raccoonforlemmy.core.commonui.modals.di
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.LocalView
+import androidx.compose.ui.platform.TextToolbar
+import com.github.diegoberaldin.raccoonforlemmy.core.commonui.modals.CustomTextToolbar
+
+@Composable
+actual fun getCustomTextToolbar(
+    onShare: () -> Unit,
+    onQuote: () -> Unit,
+): TextToolbar {
+    return CustomTextToolbar(
+        view = LocalView.current,
+        onShare = onShare,
+        onQuote = onQuote,
+    )
+}
