+++ core-commonui/modals/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/core/commonui/modals/di/Utils.kt
@@ -1,0 +1,10 @@
+package com.github.diegoberaldin.raccoonforlemmy.core.commonui.modals.di
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.TextToolbar
+
+@Composable
+expect fun getCustomTextToolbar(
+    onShare: () -> Unit,
+    onQuote: () -> Unit,
+): TextToolbar
