+++ core-commonui/modals/src/iosMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/core/commonui/modals/di/Utils.kt
@@ -1,0 +1,13 @@
+package com.github.diegoberaldin.raccoonforlemmy.core.commonui.modals.di
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.LocalTextToolbar
+import androidx.compose.ui.platform.TextToolbar
+
+@Composable
+actual fun getCustomTextToolbar(
+    onShare: () -> Unit,
+    onQuote: () -> Unit,
+): TextToolbar {
+    return LocalTextToolbar.current
+}
