+++ feature-profile/build.gradle.kts
@@ -53,6 +53,7 @@ kotlin {
                 implementation(projects.coreCommonui)
                 implementation(projects.coreCommonui.components)
                 implementation(projects.coreCommonui.lemmyui)
+                implementation(projects.coreCommonui.modals)
                 implementation(projects.corePreferences)
                 implementation(projects.coreNotifications)
                 implementation(projects.corePersistence)
