+++ feature-search/build.gradle.kts
@@ -52,6 +52,7 @@ kotlin {
                 implementation(projects.coreCommonui)
                 implementation(projects.coreCommonui.components)
                 implementation(projects.coreCommonui.lemmyui)
+                implementation(projects.coreCommonui.modals)
                 implementation(projects.coreNotifications)
 
                 implementation(projects.domainIdentity)
