+++ settings.gradle.kts
@@ -29,6 +29,7 @@ include(":core-md")
 include(":core-commonui")
 include(":core-commonui:components")
 include(":core-commonui:lemmyui")
+include(":core-commonui:modals")
 include(":core-notifications")
 include(":core-persistence")
 include(":core-navigation")
